All Bot Choice Makers, Bot bet makers and so on are ready. Started debugging. shows errors with initializing ChipController for now.

+++ High-Quality-Code-Teamwork-Project-Poker/Poker2/Core/Interfaces/IDatabase.cs
@@ -12,7 +12,7 @@
 
         int CallAmount { get; set; }
 
-        int RaiseAmount { get; set; }
+        double RaiseAmount { get; set; }
 
         int BigBlind { get; set; }
